On window resize, the window now forces the aspect ratio 4:3.

+++ ColorDefense/util.hpp
@@ -3,11 +3,11 @@
 #include "gamelogic.hpp"
 
 inline sf::Vector2i convertToCellCoords(float x, float y) {
-	return sf::Vector2i((int) x / TILEWIDTH, (int) y / TILEHEIGHT);
+	return sf::Vector2i(static_cast<int>(x / TILEWIDTH), static_cast<int>(y / TILEHEIGHT));
 }
 
 inline sf::Vector2f convertToPixelCoords(int x, int y) {
-	return sf::Vector2f(x * TILEWIDTH, y * TILEHEIGHT);
+	return sf::Vector2f(static_cast<float>(x * TILEWIDTH), static_cast<float>(y * TILEHEIGHT));
 }
 
 inline float clamp(float x, float a, float b) {
